Add browser layer when the buffer is http or https
